Move the _contruct() from AccessMethodBase.php to NodeAccess.php and remove.
